chore: transformer auto-resolve ambiguous types (#2024)
